ensure latest revision is returned for proxies and shared flows
Empirically, the Apigee proxy and shared flow APIs return the list of
revisions sorted alphabetically rather than numerically. This means, for
example, that a proxy with 10 revisions will have the revision list
returned as 1, 10, 2, 3, 4, 5, 6, 7, 8, 9. This in turn means that once
a proxy or shared flow reaches revision 9, that will continue to be its
"latest" revision until revision 99 or 999, etc.

Fix this by iterating over the list of revisions returned by the API and
selecting the maximum when parsed as an integer.

Closes #33