shared_flow_deployment: allow specifying a service account
Much as a proxy deployment can take a service account when deployed,
shared flow can do the same. Use the same implementation as used with
proxy_deployment for shared_flow_deployment.

Additionally, it appears that the Apigee API returns service accounts
prefixed with "projects/-/serviceAccounts/" when reading the resource
back and so trim that prefix from the service account returned from the
API.